refactor(PaymentConfig): defaultHeader 추가
